move Collection to src/python/pants/engine/objects.py
