+++ src/python/pants/engine/legacy/BUILD
@@ -75,6 +75,7 @@ python_library(
     'src/python/pants/build_graph',
     'src/python/pants/engine:build_files',
     'src/python/pants/engine:mapper',
+    'src/python/pants/engine:objects',
     'src/python/pants/engine:parser',
     'src/python/pants/engine:selectors',
     'src/python/pants/option',
