+++ tests/python/pants_test/engine/BUILD
@@ -156,6 +156,7 @@ python_tests(
     'src/python/pants/build_graph',
     'src/python/pants/engine:build_files',
     'src/python/pants/engine:mapper',
+    'src/python/pants/engine:objects',
     'src/python/pants/engine:struct',
     'src/python/pants/util:dirutil',
     'tests/python/pants_test/engine/examples:mapper_test',
@@ -231,4 +232,13 @@ python_library(
     'src/python/pants/engine:scheduler',
     'src/python/pants/util:dirutil',
   ]
+)
+
+python_tests(
+  name='objects',
+  sources=['test_objects.py'],
+  dependencies=[
+    'src/python/pants/engine:objects',
+    'tests/python/pants_test:test_base',
+  ],
 )
